+++ libraries/test_session_current/src/main/java/androidx/media3/session/MockPlayer.java
@@ -871,7 +871,7 @@ public class MockPlayer implements Player {
 
   @Override
   public int getMediaItemCount() {
-    throw new UnsupportedOperationException();
+    return timeline.getWindowCount();
   }
 
   @Override