+++ RELEASENOTES.md
@@ -62,6 +62,10 @@
         ([#355](https://github.com/androidx/media/issues/355)).
     *   Fix memory leak of `MediaSessionService` or `MediaLibraryService`
         ([#346](https://github.com/androidx/media/issues/346)).
+    *   Add `androidx.media3.session.MediaButtonReceiver` to enable apps to
+        implement playback resumption with media button events sent by, for
+        example, a Bluetooth headset
+        ([#167](https://github.com/androidx/media/issues/167)).
 *   UI:
     *   Add Util methods `shouldShowPlayButton` and
         `handlePlayPauseButtonAction` to write custom UI elements with a